Remove XVFB related code from the library, run xvfb for the tests.
Fixes #19.
