+++ tbselenium/test/test_env.py
@@ -47,9 +47,6 @@ class EnvironmentTest(unittest.TestCase):
     def test_tld(self):
         self.assert_py_pkg_installed('tld')
 
-    def test_xvfwrapper(self):
-        self.assert_py_pkg_installed('pyvirtualdisplay')
-
     def test_selenium(self):
         self.assert_py_pkg_installed('selenium')
